tweak secondary structure extraction
