+++ molecularnodes/io/load.py
@@ -67,68 +67,7 @@ def pdb_get_b_factors(file):
     return b_factors
 
 
-def get_secondary_structure(array, file) -> np.array:
-    """
-    Gets the secondary structure annotation that is included in mmtf files and returns it as a numerical numpy array.
 
-    Parameters:
-    -----------
-    array : numpy.array
-        The molecular coordinates array, from mmtf.get_structure()
-    file : mmtf.MMTFFile
-        The MMTF file containing the secondary structure information, from mmtf.MMTFFile.read()
-
-    Returns:
-    --------
-    atom_sse : numpy.array
-        Numerical numpy array representing the secondary structure of the molecule.
-    
-    Description:
-    ------------
-    This function uses the biotite.structure package to extract the secondary structure information from the MMTF file.
-    The resulting secondary structures are `1: Alpha Helix, 2: Beta-sheet, 3: loop`.
-    """
-    
-    from biotite.structure import spread_residue_wise
-    
-    sec_struct_codes = {
-        -1: "X",
-        0 : "I",
-        1 : "S",
-        2 : "H",
-        3 : "E",
-        4 : "G",
-        5 : "B",
-        6 : "T",
-        7 : "C"
-    }
-    
-    dssp_to_abc = {
-        "X" : 0,
-        "I" : 1, #"a",
-        "S" : 3, #"c",
-        "H" : 1, #"a",
-        "E" : 2, #"b",
-        "G" : 1, #"a",
-        "B" : 2, #"b",
-        "T" : 3, #"c",
-        "C" : 3 #"c"
-    }
-    
-    try:
-        sse = file["secStructList"]
-    except KeyError:
-        ss_int = np.full(len(array), 3)
-        print('Warning: "secStructList" field missing from MMTF file. Defaulting \
-            to "loop" for all residues.')
-    else:
-        ss_int = np.array(
-            [dssp_to_abc.get(sec_struct_codes.get(ss)) for ss in sse], 
-            dtype = int
-        )
-    atom_sse = spread_residue_wise(array, ss_int)
-    
-    return atom_sse
 
 
 def comp_secondary_structure(array):
@@ -343,14 +282,8 @@ def create_molecule(array,
     def att_sec_struct():
         if hasattr(array, "sec_struct"):
             return array.sec_struct
-        if calculate_ss or not file:
+        if calculate_ss:
             return comp_secondary_structure(array)
-        else:
-            return get_secondary_structure(array, file)
-    
-
-    
-
     
     # these are all of the attributes that will be added to the structure
     # TODO add capcity for selection of particular attributes to include / not include to potentially
